pyproject.toml
[project]
name = "linktolead"
version = "0.1.0"
description = "A CLI tool to parse LinkedIn job and company PDFs and create corresponding Deal and Company entries in HubSpot."
authors = [
    "Snowpack Data <company@snowpack-data.io>",
    "David Olodort <olodort@snowpack-data.com>",
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "pdfplumber>=0.9.0",
    "requests>=2.28.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=10.11.0",
    "questionary>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.0.267",
]

# Optional LLM dependencies (only needed if llm_enabled=true in config)
llm = [
    "llm>=0.22",
    "llm-mlx>=0.2.1",
]

[project.scripts]
linktolead = "linktolead.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
